updating accelwattch ubenchs to read in 64b input

+++ src/cuda/accelwattch-ubench/branching_benchmarks/BE_SP_FP_MUL_l32/BE_SP_FP_MUL_l32.cu
@@ -79,7 +79,7 @@ inline void __getLastCudaError(const char *errorMessage, const char *file, const
 }
 // end of CUDA Helper Functions
 
-__global__ void PowerKernal2(const float* A, const float* B, float* C, int iterations)
+__global__ void PowerKernal2(const float* A, const float* B, float* C, unsigned long long iterations)
 {
   int i = blockDim.x * blockIdx.x + threadIdx.x;
   //Do Some Computation
@@ -93,7 +93,7 @@ __global__ void PowerKernal2(const float* A, const float* B, float* C, int itera
   // Excessive Division access
   if((i%32)<=31){
     #pragma unroll 100
-    for(unsigned k=0; k<iterations;k++) {
+    for(unsigned long long k=0; k<iterations;k++) {
       Value1=I1*I2;
       Value3=I1*I2;
       Value1*=Value2;
@@ -111,13 +111,13 @@ __global__ void PowerKernal2(const float* A, const float* B, float* C, int itera
 int main(int argc, char** argv) 
 {
 
-  int iterations;
+  unsigned long long iterations;
   if (argc != 2){
   fprintf(stderr,"usage: %s #iterations\n",argv[0]);
   exit(1);
   }
   else{
-    iterations = atoi(argv[1]);
+    iterations = atoll(argv[1]);
   }
 
  printf("Power Microbenchmark with %d iterations\n",iterations);
