+++ src/cuda/accelwattch-ubench/functional_benchmarks/BE_DP_FP_MUL/BE_DP_FP_MUL.cu
@@ -87,7 +87,7 @@ inline void __getLastCudaError(const char *errorMessage, const char *file, const
 }
 
 // end of CUDA Helper Functions
-__global__ void PowerKernal2(const double* A, const double* B, double* C, int N)
+__global__ void PowerKernal2(const double* A, const double* B, double* C, unsigned long long N)
 {
     int i = blockDim.x * blockIdx.x + threadIdx.x;
     //Do Some Computation
@@ -99,7 +99,7 @@ __global__ void PowerKernal2(const double* A, const double* B, double* C, int N)
     double I2=B[i];
 #pragma unroll 100
     // Excessive Addition access
-    for(unsigned k=0; k<N;k++) {
+    for(unsigned long long k=0; k<N;k++) {
 	Value1=I1*I2;
 	Value3=I1*I2;
 	Value1*=Value2;
@@ -123,16 +123,16 @@ __global__ void PowerKernal2(const double* A, const double* B, double* C, int N)
 
 int main(int argc, char** argv)
 {
- int iterations;
+ unsigned long long iterations;
  if(argc!=2) {
    fprintf(stderr,"usage: %s #iterations\n",argv[0]);
    exit(1);
  }
  else {
-   iterations = atoi(argv[1]);
+   iterations = atoll(argv[1]);
  }
  
- printf("Power Microbenchmarks with iterations %d\n",iterations);
+ printf("Power Microbenchmarks with iterations %lld\n",iterations);
  int N = THREADS_PER_BLOCK*NUM_OF_BLOCKS;
  size_t size = N * sizeof(double);
  // Allocate input vectors h_A and h_B in host memory
