+++ src/cuda/accelwattch-ubench/functional_benchmarks/BE_HP_FP_ADD/BE_HP_FP_ADD.cu
@@ -105,7 +105,7 @@ __global__ void convertFp32ToFp16 (half *out, float *in, int n) {
 
 
 // Device code
-__global__ void PowerKernal1(const half* A, const half* B, half* C, int iterations)
+__global__ void PowerKernal1(const half* A, const half* B, half* C, unsigned long long iterations)
 {
     int i = blockDim.x * blockIdx.x + threadIdx.x;
     //Do Some Computation
@@ -117,7 +117,7 @@ __global__ void PowerKernal1(const half* A, const half* B, half* C, int iteratio
     volatile half I2=B[i];
 #pragma unroll 100
     // Excessive Addition access
-    for(int k=0; k<iterations;k++) {
+    for(unsigned long long k=0; k<iterations;k++) {
     	Value1= __hadd(I1,I2);
     	Value3= __hsub(I1,I2);
     	Value1= __hadd(Value1,Value2);
@@ -134,16 +134,16 @@ __global__ void PowerKernal1(const half* A, const half* B, half* C, int iteratio
 
 int main(int argc, char** argv)
 {
- int iterations;
+ unsigned long long iterations;
  if(argc!=2) {
    fprintf(stderr,"usage: %s #iterations\n",argv[0]);
    exit(1);
  }
  else {
-   iterations = atoi(argv[1]);
+   iterations = atoll(argv[1]);
  }
  
- printf("Power Microbenchmarks with iterations %d\n",iterations);
+ printf("Power Microbenchmarks with iterations %lld\n",iterations);
  int N = THREADS_PER_BLOCK*NUM_OF_BLOCKS;
  size_t size = N * sizeof(float);
  // Allocate input vectors h_A and h_B in host memory
