+++ src/cuda/accelwattch-ubench/functional_benchmarks/BE_HP_FP_MUL/BE_HP_FP_MUL.cu
@@ -100,7 +100,7 @@ __global__ void convertFp32ToFp16 (half *out, float *in, int n) {
    }
 }
 // end of CUDA Helper Functions
-__global__ void PowerKernal2(const half* A, const half* B, half* C, int N)
+__global__ void PowerKernal2(const half* A, const half* B, half* C, unsigned long long N)
 {
     int i = blockDim.x * blockIdx.x + threadIdx.x;
     //Do Some Computation
@@ -112,7 +112,7 @@ __global__ void PowerKernal2(const half* A, const half* B, half* C, int N)
     half I2=B[i];
 #pragma unroll 100
     // Excessive Addition access
-    for(unsigned k=0; k<N;k++) {
+    for(unsigned long long k=0; k<N;k++) {
 	Value1= __hmul(I1,I2);
 	Value3= __hmul(I1,I2);
 	Value1= __hmul(Value1,Value2);
@@ -136,16 +136,16 @@ __global__ void PowerKernal2(const half* A, const half* B, half* C, int N)
 
 int main(int argc, char** argv)
 {
- int iterations;
+ unsigned long long iterations;
  if(argc!=2) {
    fprintf(stderr,"usage: %s #iterations\n",argv[0]);
    exit(1);
  }
  else {
-   iterations = atoi(argv[1]);
+   iterations = atoll(argv[1]);
  }
  
- printf("Power Microbenchmarks with iterations %d\n",iterations);
+ printf("Power Microbenchmarks with iterations %lld\n",iterations);
  int N = THREADS_PER_BLOCK*NUM_OF_BLOCKS;
  size_t size = N * sizeof(float);
  // Allocate input vectors h_A and h_B in host memory
