+++ src/cuda/accelwattch-ubench/functional_benchmarks/BE_SP_INT_DIV/BE_SP_INT_DIV.cu
@@ -93,7 +93,7 @@ inline void __getLastCudaError(const char *errorMessage, const char *file, const
 
 
 
-__global__ void PowerKernal4(const unsigned* A, const unsigned* B, unsigned* C, int iterations)
+__global__ void PowerKernal4(const unsigned* A, const unsigned* B, unsigned* C, unsigned long long iterations)
 {
     int i = blockDim.x * blockIdx.x + threadIdx.x;
     //Do Some Computation
@@ -108,7 +108,7 @@ __global__ void PowerKernal4(const unsigned* A, const unsigned* B, unsigned* C, 
     __syncthreads();
     #pragma unroll 100
    // Excessive Mod/Div Operations
-    for(unsigned k=0; k<iterations;k++) {
+    for(unsigned long long k=0; k<iterations;k++) {
 
     	Value1=I1/(I2+1);
     	Value2=Value1/(I2+1);
@@ -140,15 +140,15 @@ __global__ void PowerKernal4(const unsigned* A, const unsigned* B, unsigned* C, 
 
 int main(int argc, char** argv)
 {
- int iterations;
+ unsigned long long iterations;
  if(argc!=2) {
    fprintf(stderr,"usage: %s #iterations\n",argv[0]);
    exit(1);
  }
  else {
-   iterations = atoi(argv[1]);
+   iterations = atoll(argv[1]);
  }
- printf("Power Microbenchmarks with iterations %d\n",iterations);
+ printf("Power Microbenchmarks with iterations %lld\n",iterations);
  int N = THREADS_PER_BLOCK*NUM_OF_BLOCKS;
  size_t size = N * sizeof(unsigned);
  // Allocate input vectors h_A and h_B in host memory
