+++ src/cuda/accelwattch-ubench/functional_benchmarks/MOV/MOV.cu
@@ -93,7 +93,7 @@ inline void __getLastCudaError(const char *errorMessage, const char *file, const
 
 
 
-__global__ void PowerKernal2(const unsigned* A, const unsigned* B, unsigned* C, int N)
+__global__ void PowerKernal2(const unsigned* A, const unsigned* B, unsigned* C, unsigned long long N)
 {
     int i = blockDim.x * blockIdx.x + threadIdx.x;
     //Do Some Computation
@@ -104,7 +104,7 @@ __global__ void PowerKernal2(const unsigned* A, const unsigned* B, unsigned* C, 
     unsigned Value1;
     #pragma unroll 100
     //Excessive Logical Unit access
-    for(unsigned k=0; k<N;k++) {
+    for(unsigned long long k=0; k<N;k++) {
     // BLOCK-0 (For instruction size of 16 bytes for Volta
       __asm volatile (
           "\n\tmov.u32 %0, %3;"
@@ -126,15 +126,15 @@ __global__ void PowerKernal2(const unsigned* A, const unsigned* B, unsigned* C, 
 
 int main(int argc, char** argv)
 {
- int iterations;
+ unsigned long long iterations;
  if(argc!=2) {
    fprintf(stderr,"usage: %s #iterations\n",argv[0]);
  }
  else {
-   iterations = atoi(argv[1]);
+   iterations = atoll(argv[1]);
  }
  
- printf("Power Microbenchmarks with iterations %d\n",iterations);
+ printf("Power Microbenchmarks with iterations %lld\n",iterations);
  int N = THREADS_PER_BLOCK*NUM_OF_BLOCKS;
  size_t size = N * sizeof(unsigned);
  // Allocate input vectors h_A and h_B in host memory
