+++ src/cuda/accelwattch-ubench/memories_benchmarks/BE_L2D_MISS/BE_L2D_MISS.cu
@@ -95,7 +95,7 @@ inline void __getLastCudaError(const char *errorMessage, const char *file, const
 
 
 
-__global__ void PowerKernal2( unsigned* A, unsigned* B, int N)
+__global__ void PowerKernal2( unsigned* A, unsigned* B, unsigned long long N)
 {
     int i = blockDim.x * blockIdx.x + threadIdx.x;
 
@@ -105,7 +105,7 @@ __global__ void PowerKernal2( unsigned* A, unsigned* B, int N)
 	//unsigned sum_value = 0;
 	#pragma unroll 100
 
-    for(unsigned k=0; k<N;k++) {
+    for(unsigned long long k=0; k<N;k++) {
     	__asm volatile(
     		"ld.global.cv.u32 %0, [%1];" 
     		: "=r"(load_value) : "l"((unsigned long)(A+i))
@@ -125,16 +125,16 @@ __global__ void PowerKernal2( unsigned* A, unsigned* B, int N)
 
 int main(int argc, char** argv)
 {
- int iterations;
+ unsigned long long iterations;
  if(argc!=2) {
    fprintf(stderr,"usage: %s #iterations\n",argv[0]);
    exit(1);
  }
  else {
-   iterations = atoi(argv[1]);
+   iterations = atoll(argv[1]);
  }
  
- printf("Power Microbenchmarks with iterations %d\n",iterations);
+ printf("Power Microbenchmarks with iterations %lld\n",iterations);
  int size_per_sm = (LINE_SIZE*ASSOC*SETS); //131072
  int N = THREADS_PER_BLOCK*NUM_OF_BLOCKS;
